+++ src/places/index.ts
@@ -33,6 +33,8 @@ export {
 	isBuiltInPlaceTypeCategory,
 	getAllPlaceTypeCategories,
 	reorderPlaceTypeCategories,
+	nextHierarchyLevelForCategory,
+	reorderTypeWithinCategory,
 	getCategoryHierarchyLevelRange,
 	getCategoryForHierarchyLevel
 } from './constants/default-place-types';